backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
websockets==14.1
pyyaml==6.0.2
pydantic==2.10.4
python-multipart==0.0.20
